python_service/requirements.txt
psycopg2-binary
watchdog
pandas
scikit-learn